feat: separated base functionality into their respective classes
started to comment why the order of some things matter (specially
imports)
